requirements.txt
# PyTorch (install separately with CUDA support)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.65.0

# Deep Learning
torchvision>=0.15.0

# Image quality metrics
scikit-image>=0.21.0
scipy>=1.10.0
pytorch-fid>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Jupyter notebook support
# jupyter>=1.0.0
# ipykernel>=6.0.0
